requirements.txt
fastapi[standard]
uvicorn
python-multipart
supabase
pyjwt
requests